pyproject.toml
[tool.poetry]
name = "finmatcher"
version = "0.1.0"
description = ""
package-mode = false
authors = ["qasim0900 <67171154+qasim0900@users.noreply.github.com>"]

readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
pandas = ">=3.0.1,<4.0.0"
fuzzywuzzy = ">=0.18.0,<0.19.0"
python-levenshtein = ">=0.27.3,<0.28.0"
google-api-python-client = ">=2.190.0,<3.0.0"
google-auth-oauthlib = ">=1.2.4,<2.0.0"
google-auth = ">=2.48.0,<3.0.0"
pymupdf = ">=1.27.1,<2.0.0"
openpyxl = ">=3.1.5,<4.0.0"
python-dotenv = ">=1.2.1,<2.0.0"
python-dateutil = ">=2.9.0.post0,<3.0.0"
scikit-learn = ">=1.8.0,<2.0.0"
joblib = ">=1.5.3,<2.0.0"
openai = ">=1.0.0,<2.0.0"
numpy = ">=2.0.0,<3.0.0"
aiofiles = "^25.1.0"
pytesseract = ">=0.3.10,<0.4.0"
pillow = ">=10.0.0,<11.0.0"
scipy = ">=1.11.0,<2.0.0"

[tool.poetry.group.test.dependencies]
pytest = ">=8.0.0,<9.0.0"
pytest-cov = ">=4.1.0,<5.0.0"
pytest-mock = ">=3.12.0,<4.0.0"
pytest-asyncio = ">=0.23.0,<0.24.0"
pytest-xdist = ">=3.5.0,<4.0.0"
pytest-timeout = ">=2.2.0,<3.0.0"
hypothesis = ">=6.98.0,<7.0.0"
coverage = ">=7.4.0,<8.0.0"
faker = ">=22.0.0,<23.0.0"
freezegun = ">=1.4.0,<2.0.0"
responses = ">=0.24.0,<0.25.0"
pytest-benchmark = ">=4.0.0,<5.0.0"
memory-profiler = ">=0.61.0,<0.62.0"
pytest-html = ">=4.1.0,<5.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"